+++ .env.sample
@@ -1,0 +1,5 @@
+NODE_ENV=
+PORT=
+HOST=
+PROVIDER_WS=
+PROVIDER_HTTP=
