+++ src/lib/helpers.js
@@ -1,4 +1,12 @@
-const { TRANCHE_DURATION_DAYS, MAX_ACTIVE_TRANCHES, PRICE_CHANGE_PER_DAY } = require('./constants');
+const {
+  TRANCHE_DURATION_DAYS,
+  MAX_ACTIVE_TRANCHES,
+  PRICE_CHANGE_PER_DAY,
+  NXM_PER_ALLOCATION_UNIT,
+  TARGET_PRICE_DENOMINATOR,
+  SURGE_THRESHOLD_RATIO,
+  SURGE_PRICE_RATIO,
+} = require('./constants');
 const { BigNumber } = require('ethers');
 
 // offest in days
@@ -37,8 +45,54 @@ function sortPools([, a], [, b]) {
   return 0;
 }
 
+// PREMIUM CALCULATIONS
+function calculatePremium(
+  coverAmount,
+  period,
+  targetPrice,
+  bumpedPrice,
+  secondsSinceLastUpdate,
+  initialCapacityUsed,
+  totalCapacity,
+) {
+  const priceDrop = Math.floor((PRICE_CHANGE_PER_DAY * secondsSinceLastUpdate) / 86_400);
+  const basePrice = Math.max(targetPrice, bumpedPrice - priceDrop);
+
+  // const basePremium = (coverAmount * NXM_PER_ALLOCATION_UNIT * basePrice) / TARGET_PRICE_DENOMINATOR;
+  const basePremium = (coverAmount * basePrice * NXM_PER_ALLOCATION_UNIT) / TARGET_PRICE_DENOMINATOR;
+  const finalCapacityUsed = initialCapacityUsed.toNumber() + coverAmount;
+
+  const surgeStartPoint = totalCapacity.toNumber() * SURGE_THRESHOLD_RATIO;
+
+  if (surgeStartPoint >= finalCapacityUsed) {
+    return Math.floor((basePremium * period) / (365 * 86_400));
+  }
+
+  const amountOnSurgeSkipped = initialCapacityUsed - surgeStartPoint > 0 ? initialCapacityUsed - surgeStartPoint : 0;
+
+  const amountOnSurge = finalCapacityUsed - surgeStartPoint;
+  let surgePremium = 0;
+  if (amountOnSurge > 0) {
+    surgePremium = calculateSurgePremium(amountOnSurge, totalCapacity, amountOnSurgeSkipped);
+  }
+  return Math.floor(((basePremium + surgePremium) * period) / (365 * 86_400));
+}
+
+function calculateSurgePremium(amountOnSurge, totalCapacity, amountOnSurgeSkipped = 0) {
+  let surgePremium = (amountOnSurge * SURGE_PRICE_RATIO * (amountOnSurge / totalCapacity)) / 2;
+
+  if (amountOnSurgeSkipped > 0) {
+    surgePremium =
+      surgePremium - (amountOnSurgeSkipped * SURGE_PRICE_RATIO * (amountOnSurgeSkipped / totalCapacity)) / 2;
+  }
+  // amountOnSurge has two decimals
+  // dividing by ALLOCATION_UNITS_PER_NXM (=100) to normalize the result
+  return surgePremium / NXM_PER_ALLOCATION_UNIT;
+}
+
 module.exports = {
   calculateTranche,
   calculateCapacities,
   sortPools,
+  calculatePremium,
 };
